KS-2-1 mean value refactor numberOfStudents

+++ src/main/kotlin/lesson2/lesson2-task1.kt
@@ -3,7 +3,7 @@ package org.example.lesson2
 fun main() {
 
     val numberOfStudents = 4
-    val meanEnglishGrade: Float = (3.0f + 4.0f + 3.0f + 5.0f) / numberOfStudents.toFloat()
+    val meanEnglishGrade:Float = (3.0f + 4.0f + 3.0f + 5.0f) / numberOfStudents
 
     println(meanEnglishGrade)
 }
